app/gpu_module/requirements.txt
numpy
torchmeta
